# Allow upstream for Slice on single axis