--- a/onnxruntime/core/optimizer/compute_optimizer/upstream_gather.cc
+++ b/onnxruntime/core/optimizer/compute_optimizer/upstream_gather.cc
@@ -345,6 +345,56 @@ std::optional<SliceInfo> IsSupportedShrunkenGather(Graph& graph, Node& node,
   return SliceInfo(graph, &node, false /*is_slice_scalar*/, "axis", axis, true);
 }
 
+/**
+ * @brief Check if the Slice node can be up-streamed to the previous node.
+ *
+ * If "Slice" node is operating on one single axis, then it is supported.
+ * @return std::optional<SliceInfo>
+ */
+std::optional<SliceInfo> IsSupportedSlice(Graph& graph, Node& node,
+                                          const InlinedHashSet<std::string_view>&
+                                              compatible_execution_providers,
+                                          const logging::Logger& logger) {
+  if (!graph_utils::IsSupportedOptypeVersionAndDomain(node, "Slice", {10, 11, 13}) ||
+      !graph_utils::IsSupportedProvider(node, compatible_execution_providers)) {
+    return std::nullopt;
+  }
+
+  const NodeArg* data_input = node.InputDefs()[0];
+  const NodeArg* starts_input = node.InputDefs()[1];
+  const NodeArg* ends_input = node.InputDefs()[2];
+  const NodeArg* axes_input = node.InputDefs().size() > 3 ? node.InputDefs()[3] : nullptr;
+
+  if (data_input->Shape() == nullptr || starts_input->Shape() == nullptr || ends_input->Shape() == nullptr ||
+      (axes_input && axes_input->Shape() == nullptr)) {
+    LOG_DEBUG_INFO(logger, "Skip Slice node " + node.Name() + " due to undefined shape.");
+    return std::nullopt;
+  }
+
+  // Make sure starts/ends/axes/steps are all 1D tensors, since we only support single dimension slicing.
+  if (starts_input->Shape()->dim_size() != 1 || ends_input->Shape()->dim_size() != 1 ||
+      (axes_input && axes_input->Shape()->dim_size() != 1)) {
+    LOG_DEBUG_INFO(logger, "Skip Slice node " + node.Name() + " due to unsupported dim size: " +
+                               std::to_string(starts_input->Shape()->dim_size()) + ", " +
+                               std::to_string(ends_input->Shape()->dim_size()) + ", " +
+                               std::to_string(axes_input ? axes_input->Shape()->dim_size() : 0));
+    return std::nullopt;
+  }
+
+  // Try to parse the axes value.
+  InlinedVector<int64_t> axes_values;
+  if (axes_input && !(optimizer_utils::AppendTensorFromInitializer(graph, *axes_input, axes_values, true) &&
+                      axes_values.size() == 1)) {
+    return std::nullopt;
+  }
+
+  int axis = axes_values[0];
+  if (axis < 0)
+    axis += data_input->Shape()->dim_size();
+
+  return SliceInfo(graph, &node, false /*is_slice_scalar*/, "axis", axis, true);
+}
+
 }  // namespace
 
 std::optional<SliceInfo> UpStreamGatherGraphTransformer::IsSupportedForUpstream(
@@ -357,6 +407,9 @@ std::optional<SliceInfo> UpStreamGatherGraphTransformer::IsSupportedForUpstream(
   }
   if (!gather_info.has_value()) {
     gather_info = IsSupportedShrunkenGather(graph, node, GetCompatibleExecutionProviders(), logger);
+  }
+  if (!gather_info.has_value() && true) {
+    gather_info = IsSupportedSlice(graph, node, GetCompatibleExecutionProviders(), logger);
   }
   return gather_info;
 }